pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "extracosmic-commons"
version = "0.1.0"
description = "Privacy-first, cooperatively owned scholarly research platform with citation-aware multilingual RAG"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Douglas Locklin", email = "douglas.locklin@gmail.com"},
]
dependencies = [
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.7",
    "click>=8.0",
    "pypdf>=4.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
ec = "extracosmic_commons.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
